refactor: replace all client-common imports
