--- packages/analytics-client-common/jest.config.js
@@ -1,10 +1,0 @@
-const baseConfig = require('../../jest.config.js');
-const package = require('./package');
-
-module.exports = {
-  ...baseConfig,
-  displayName: package.name,
-  rootDir: '.',
-  testEnvironment: 'jsdom',
-  coveragePathIgnorePatterns: ['index.ts', 'global-scope.ts'],
-};
