--- packages/analytics-client-common/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "@amplitude/analytics-client-common",
-  "version": "2.4.58",
-  "description": "",
-  "author": "Amplitude Inc",
-  "homepage": "https://github.com/amplitude/Amplitude-TypeScript",
-  "license": "MIT",
-  "main": "lib/cjs/index.js",
-  "module": "lib/esm/index.js",
-  "types": "lib/esm/index.d.ts",
-  "sideEffects": false,
-  "publishConfig": {
-    "access": "public",
-    "tag": "latest"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/amplitude/Amplitude-TypeScript.git"
-  },
-  "scripts": {
-    "build": "pnpm build:es5 && pnpm build:esm",
-    "build:es5": "tsc -p ./tsconfig.es5.json",
-    "build:esm": "tsc -p ./tsconfig.esm.json",
-    "watch": "tsc -p ./tsconfig.esm.json --watch",
-    "clean": "rimraf node_modules lib coverage",
-    "fix": "pnpm fix:eslint & pnpm fix:prettier",
-    "fix:eslint": "eslint '{src,test}/**/*.ts' --fix",
-    "fix:prettier": "prettier --write \"{src,test}/**/*.ts\"",
-    "lint": "pnpm lint:eslint & pnpm lint:prettier",
-    "lint:eslint": "eslint '{src,test}/**/*.ts'",
-    "lint:prettier": "prettier --check \"{src,test}/**/*.ts\"",
-    "test": "jest",
-    "typecheck": "tsc -p ./tsconfig.json"
-  },
-  "bugs": {
-    "url": "https://github.com/amplitude/Amplitude-TypeScript/issues"
-  },
-  "dependencies": {
-    "@amplitude/analytics-connector": "^1.4.8",
-    "@amplitude/analytics-core": "workspace:*",
-    "@amplitude/analytics-types": "workspace:*",
-    "tslib": "^2.4.1"
-  },
-  "files": [
-    "lib"
-  ]
-}
