--- packages/analytics-client-common/src/session.ts
@@ -1,6 +1,0 @@
-export const isNewSession = (sessionTimeout: number, lastEventTime: number = Date.now()): boolean => {
-  const currentTime = Date.now();
-  const timeSinceLastEvent = currentTime - lastEventTime;
-
-  return timeSinceLastEvent > sessionTimeout;
-};
