--- packages/analytics-client-common/src/types/global.ts
@@ -1,3 +1,0 @@
-/* eslint-disable no-restricted-globals */
-// eslint-disable-next-line no-var
-export var ampIntegrationContext: typeof globalThis;
