--- packages/analytics-client-common/test/helpers/constants.ts
@@ -1,1 +1,0 @@
-export const API_KEY = 'apiKey';
