--- packages/analytics-client-common/test/transports/fetch.test.ts
@@ -1,33 +1,0 @@
-import { FetchTransport } from '../../src/transports/fetch';
-import { Status } from '@amplitude/analytics-types';
-import 'isomorphic-fetch';
-
-describe('fetch', () => {
-  describe('send', () => {
-    test.each([
-      ['{}'], // ideally response body should be json format to an application/json request
-      [''], // test the edge case where response body is non-json format
-      ['<'],
-    ])('should resolve with response', async (body) => {
-      const transport = new FetchTransport();
-      const url = 'http://localhost:3000';
-      const payload = {
-        api_key: '',
-        events: [],
-      };
-      const result = {
-        statusCode: 200,
-        status: Status.Success as const,
-        body: {
-          eventsIngested: 0,
-          payloadSizeBytes: 0,
-          serverUploadTime: 0,
-        },
-      };
-      jest.spyOn(window, 'fetch').mockReturnValueOnce(Promise.resolve(new Response(body)));
-      jest.spyOn(transport, 'buildResponse').mockReturnValueOnce(result);
-      const response = await transport.send(url, payload);
-      expect(response).toEqual(result);
-    });
-  });
-});
