--- packages/analytics-client-common/tsconfig.es5.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": ["src/**/*"],
-  "compilerOptions": {
-    "module": "commonjs",
-    "noEmit": false,
-    "outDir": "lib/cjs",
-    "rootDir": "./src"
-  }
-}
