--- packages/analytics-client-common/tsconfig.esm.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": ["src/**/*"],
-  "compilerOptions": {
-    "module": "es6",
-    "noEmit": false,
-    "outDir": "lib/esm",
-    "rootDir": "./src"
-  }
-}
