--- packages/analytics-client-common/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["src/**/*", "test/**/*"],
-  "compilerOptions": {
-    "baseUrl": ".",
-    "esModuleInterop": true,
-    "lib": ["dom"],
-    "noEmit": true,
-    "rootDir": ".",
-  }
-}
