+++ packages/plugin-session-replay-browser/package.json
@@ -39,7 +39,6 @@
     "url": "https://github.com/amplitude/Amplitude-TypeScript/issues"
   },
   "dependencies": {
-    "@amplitude/analytics-client-common": "workspace:*",
     "@amplitude/analytics-core": "workspace:*",
     "@amplitude/analytics-types": "workspace:*",
     "@amplitude/rrweb-plugin-console-record": "2.0.0-alpha.40",
