+++ packages/plugin-web-attribution-browser/src/constants.ts
@@ -1,0 +1,23 @@
+import type { Campaign } from '@amplitude/analytics-core';
+
+export const BASE_CAMPAIGN: Campaign = {
+  utm_campaign: undefined,
+  utm_content: undefined,
+  utm_id: undefined,
+  utm_medium: undefined,
+  utm_source: undefined,
+  utm_term: undefined,
+  referrer: undefined,
+  referring_domain: undefined,
+  dclid: undefined,
+  gbraid: undefined,
+  gclid: undefined,
+  fbclid: undefined,
+  ko_click_id: undefined,
+  li_fat_id: undefined,
+  msclkid: undefined,
+  rdt_cid: undefined,
+  ttclid: undefined,
+  twclid: undefined,
+  wbraid: undefined,
+};
