+++ packages/plugin-web-attribution-browser/test/web-attribution.test.ts
@@ -1,8 +1,16 @@
 import { createInstance } from '@amplitude/analytics-browser';
-import { BASE_CAMPAIGN, CampaignParser, CookieStorage, FetchTransport } from '@amplitude/analytics-client-common';
 import { webAttributionPlugin } from '../src/web-attribution';
-import * as helpers from '@amplitude/analytics-client-common';
-import { BrowserConfig, LogLevel, Logger, UUID } from '@amplitude/analytics-core';
+import * as helpers from '../src/helpers';
+import {
+  BASE_CAMPAIGN,
+  BrowserConfig,
+  CampaignParser,
+  CookieStorage,
+  FetchTransport,
+  LogLevel,
+  Logger,
+  UUID,
+} from '@amplitude/analytics-core';
 
 describe('webAttributionPlugin', () => {
   const mockConfig: BrowserConfig = {
