+++ packages/session-replay-browser/package.json
@@ -40,7 +40,6 @@
     "url": "https://github.com/amplitude/Amplitude-TypeScript/issues"
   },
   "dependencies": {
-    "@amplitude/analytics-client-common": "workspace:*",
     "@amplitude/analytics-core": "workspace:*",
     "@amplitude/analytics-types": "workspace:*",
     "@amplitude/experiment-core": "0.7.2",
