+++ scripts/README.md
@@ -8,11 +8,10 @@ Runs in CI to prevent new dependencies on deprecated packages.
 
 **Deprecated packages:**
 - `@amplitude/analytics-types`
-- `@amplitude/analytics-client-common`
 - `@amplitude/analytics-remote-config`
 
 **Replacements:**
-- For `@amplitude/analytics-types` and `@amplitude/analytics-client-common`: Use `@amplitude/analytics-core`
+- For `@amplitude/analytics-types`: Use `@amplitude/analytics-core`
 - For `@amplitude/analytics-remote-config`: Use the new remote config client in `@amplitude/analytics-core`
 
 **How it works:**
